// services/document-evidence/Generated/frontend-app/src/generated-ts/staffpositions/types/CreateStaffPositionRequest.generated.ts
// ------------------------------------------------------------------------------
// GENERATED FILE - source: staffpositions/types.json
// DO NOT EDIT MANUALLY.
// ------------------------------------------------------------------------------

export interface CreateStaffPositionRequest {
  "code": string;
  "description"?: string | null;
  "isActive": boolean;
  "name": string;
  "parentPositionId"?: string | null;
}
export const CreateStaffPositionRequestTypeKey = "LogicLens.DocumentEvidence.Generated.Api.Contracts.CreateStaffPositionRequest" as const;